+++ src/Application/UseCases/Tokenables/UpsertGrantsForTokenableUseCase.php
@@ -5,14 +5,14 @@ declare(strict_types=1);
 namespace N3XT0R\LaravelPassportAuthorizationCore\Application\UseCases\Tokenables;
 
 use Illuminate\Contracts\Auth\Authenticatable;
-use N3XT0R\LaravelPassportAuthorizationCore\Events\Tokenable\TokenableGrantsAssignedEvent;
+use N3XT0R\LaravelPassportAuthorizationCore\Events\Tokenable\TokenableGrantUpsertedEvent;
 use N3XT0R\LaravelPassportAuthorizationCore\Services\GrantService;
 use N3XT0R\LaravelPassportAuthorizationCore\Support\Resolver\GrantableTokenableResolver;
 
 /**
- * Use case to assign grants to a tokenable entity.
+ * Use case to upsert grants for a tokenable entity.
  */
-readonly class AssignGrantsToTokenableUseCase
+readonly class UpsertGrantsForTokenableUseCase
 {
     public function __construct(
         protected GrantableTokenableResolver $grantableTokenableResolver,
@@ -28,14 +28,14 @@ readonly class AssignGrantsToTokenableUseCase
     ): void {
         $context = $this->grantableTokenableResolver->resolve($ownerId, $contextClientId);
 
-        $this->grantService->giveGrantsToTokenable(
+        $this->grantService->upsertGrantsForTokenable(
             tokenable: $context->tokenable,
             scopes: $scopes,
             actor: $actor,
             contextClient: $context->contextClient
         );
 
-        TokenableGrantsAssignedEvent::dispatch(
+        TokenableGrantUpsertedEvent::dispatch(
             $context->tokenable,
             $scopes,
             $context->contextClient,