+++ src/Exceptions/Domain/ActiveClientNotExistsException.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace N3XT0R\LaravelPassportAuthorizationCore\Exceptions\Domain;
+
+use DomainException;
+
+class ActiveClientNotExistsException extends DomainException implements DomainExceptionInterface
+{
+    public function __construct(string|int $id)
+    {
+        parent::__construct(
+            sprintf('No active OAuth client with id "%s" exists.', $id)
+        );
+    }
+}